python/requirements.txt
# IMG AI Studio - Requirements
# RTX 5060 Ti 16GB 최적화

# PyTorch with CUDA 12.x (RTX 50시리즈 지원)
--extra-index-url https://download.pytorch.org/whl/cu124
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0

# Diffusers (Stable Diffusion)
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# ControlNet & IP-Adapter
controlnet-aux>=0.0.7

# Optimization
xformers>=0.0.27
bitsandbytes>=0.43.0

# Utilities
tqdm>=4.66.0
huggingface-hub>=0.23.0
omegaconf>=2.3.0
einops>=0.7.0
